agent-core/pyproject.toml
[project]
name = "agent-core"
version = "0.1.0"
description = "Dynamic LangGraph-based runtime for the centralized agent platform."
requires-python = ">=3.11"
authors = [
  { name = "Aboubakar Moussa", email = "abakamousa@gmail.com" },
]
dependencies = [
  "azure-cosmos==4.9.0",
  "azure-identity==1.24.0",
  "fastapi==0.116.1",
  "httpx==0.28.1",
  "langgraph==0.6.6",
  "mlflow==2.20.3",
  "pydantic==2.11.7",
  "pyjwt[crypto]==2.10.1",
  "pyyaml==6.0.3",
  "uvicorn[standard]==0.35.0",
]

[dependency-groups]
dev = [
  "mypy==1.17.1",
  "pytest==8.4.1",
  "ruff==0.12.11",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
pretty = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra"
